+++ CHANGELOG.md
@@ -1,18 +1,22 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.3.0] 2021-01-04
+### Added
+- Verbose Logging for send and receive.
+
 ## [1.2.0] 2021-01-03
-### Changes
+### Added
 - New node for registration via voice call.
 - Save all account info in sub directory "signal" to avoid conflicts. Reregistration is neccessary.
 
 ## [1.1.0] 2020-12-31
-### Changes
+### Added
  - Documentation.
  - Registration data is stored in the user settings directory.
 
 ## [1.0.0] 2020-12-30
-### Changes
+### Added
  - Initial version.
 
 **Note:** The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).